requirements.txt
Django>=5.0.0,<5.1
djangorestframework>=3.14.0,<3.15
psycopg2 >=2.9,<3 # Postgres
drf-spectacular >=0.27, <0.28